Update import from message_infrastructure: 1. Move message_interface_enum.py to message_infrastructure 2. Update import syntax 3. Others: change multiprocessing init
