+++ src/lava/magma/compiler/builders/channel_builder.py
@@ -15,7 +15,7 @@ from lava.magma.compiler.channels.interfaces import (
     ChannelType,
 )
 from lava.magma.compiler.utils import PortInitializer
-from lava.magma.runtime.message_infrastructure.message_infrastructure \
+from message_infrastructure \
     .message_infrastructure_interface import (MessageInfrastructureInterface)
 
 if ty.TYPE_CHECKING:
