+++ src/lava/magma/runtime/message_infrastructure/message_infrastructure/message_interface_enum.py
